requirements.txt
geopy==2.0.0
de-core-news-md
mysql-connector-python==8.0.18
selenium
spacy
